+++ scripts/classes/thaumaturge.js
@@ -13,8 +13,20 @@ export const THAUMATURGE = {
   },
   skillIncreaseSchedule: [3, 5, 7, 9, 11, 13, 15, 17, 19],
   abilityBoostSchedule: [5, 10, 15, 20],
+  automaticLoreProficiencies: [
+    { level: 1, skill: 'esoteric-lore', name: 'Esoteric Lore', rank: 1 },
+    { level: 3, skill: 'esoteric-lore', name: 'Esoteric Lore', rank: 2 },
+    { level: 7, skill: 'esoteric-lore', name: 'Esoteric Lore', rank: 3 },
+    { level: 15, skill: 'esoteric-lore', name: 'Esoteric Lore', rank: 4 },
+  ],
 
   classFeatures: [
+    {
+      level: 1,
+      name: 'Exploit Vulnerability',
+      key: 'exploit-vulnerability',
+      aliases: ['mortal-weakness', 'personal-antithesis'],
+    },
     { level: 3, name: 'Reflex Expertise', key: 'reflex-expertise' },
     { level: 5, name: 'Second Implement', key: 'second-implement' },
     { level: 5, name: 'Weapon Expertise', key: 'weapon-expertise' },